xử lý lại login, logout. thêm gửi mail feedback, thông báo đẩy (demo 1 phút)

+++ app/src/main/java/com/example/proj_moneymanager/Object/UserInformation.java
@@ -12,14 +12,14 @@ import com.example.proj_moneymanager.database.DbHelper;
 import java.io.Serializable;
 
 public class UserInformation implements Serializable {
-    int UserID;
+    Long UserID;
     String FullName, UserName, PassWord, Email, PhoneNumber;
 
     public UserInformation() {
         // Constructor mặc định
     }
 
-    public UserInformation(int userID, String fullName, String userName, String passWord, String email, String phoneNumber) {
+    public UserInformation(long userID, String fullName, String userName, String passWord, String email, String phoneNumber) {
         UserID = userID;
         FullName = fullName;
         UserName = userName;
@@ -28,11 +28,11 @@ public class UserInformation implements Serializable {
         PhoneNumber = phoneNumber;
     }
 
-    public int getUserID() {
+    public long getUserID() {
         return UserID;
     }
 
-    public void setUserID(int userID) {
+    public void setUserID(long userID) {
         UserID = userID;
     }
